requirements.txt
Flask
python-docx
PyMuPDF
gunicorn
pytz